fix fc unitest
