+++ python/paddle/fluid/tests/unittests/ir/inference/test_trt_fc_fuse_pass.py
@@ -20,6 +20,7 @@ from inference_pass_test import InferencePassTest
 import paddle.fluid as fluid
 import paddle.fluid.core as core
 from paddle.fluid.core import AnalysisConfig
+from paddle.fluid.core import PassVersionChecker
 
 
 class FCFusePassTRTTest(InferencePassTest):
@@ -281,6 +282,49 @@ class FCFusePassTRTDynamicDims4Cols3Test(InferencePassTest):
             use_gpu.append(True)
         for i in range(len(use_gpu)):
             self.check_output_with_option(use_gpu[i])
+
+
+class FCQuantDequantTest(InferencePassTest):
+    def setUp(self):
+        with fluid.program_guard(self.main_program, self.startup_program):
+            data = fluid.data(
+                name="data", shape=[-1, 3, 32, 32], dtype="float32")
+            param_attr = fluid.ParamAttr(
+                initializer=fluid.initializer.Constant(value=0.0),
+                trainable=False)
+            quantized_op_out = self.append_quantized_op(data, param_attr)
+            relu_out = fluid.layers.relu(quantized_op_out)
+        self.set_quant_pattern()
+
+        self.feeds = {
+            "data": np.random.random([1, 3, 32, 32]).astype("float32"),
+        }
+        self.enable_trt = True
+        self.trt_parameters = FCQuantDequantTest.TensorRTParam(
+            1 << 30, 32, 0, AnalysisConfig.Precision.Int8, False, False)
+        self.fetch_list = [relu_out]
+
+    def append_quantized_op(self, x, param_attr):
+        return fluid.layers.fc(x,
+                               size=100,
+                               num_flatten_dims=3,
+                               param_attr=param_attr,
+                               bias_attr=False,
+                               act=None)
+
+    def set_quant_pattern(self):
+        self.activation_quant_type = 'moving_average_abs_max'
+        self.weight_quant_type = 'abs_max'
+        self.quantized_op_type = 'mul'
+        self.channels = 1
+
+    def test_check_output(self):
+        if core.is_compiled_with_cuda():
+            use_gpu = True
+            self.check_output_with_option(use_gpu, flatten=False, quant=True)
+            self.assertTrue(
+                PassVersionChecker.IsCompatible(
+                    'quant_conv2d_dequant_fuse_pass'))
 
 
 if __name__ == "__main__":
