+++ python/paddle/fluid/tests/unittests/ir/inference/test_trt_quant_conv2d_dequant_fuse_pass.py
@@ -24,109 +24,23 @@ from paddle.fluid.contrib.slim.quantization import QuantizationFreezePass
 from paddle.fluid.core import AnalysisConfig
 
 
-class Quant_Fc1_DequantTest(InferencePassTest):
+class QuantDequantTest(InferencePassTest):
     def setUp(self):
         with fluid.program_guard(self.main_program, self.startup_program):
             data = fluid.data(
                 name="data", shape=[-1, 3, 32, 32], dtype="float32")
             param_attr = fluid.ParamAttr(
-                initializer=fluid.initializer.Constant(value=1.0),
+                initializer=fluid.initializer.Constant(value=0.0),
                 trainable=False)
             quantized_op_out = self.append_quantized_op(data, param_attr)
             relu_out = fluid.layers.relu(quantized_op_out)
-            self.set_quant_pattern()
+        self.set_quant_pattern()
 
         self.feeds = {
             "data": np.random.random([1, 3, 32, 32]).astype("float32"),
         }
         self.enable_trt = True
-        self.trt_parameters = Quant_Fc1_DequantTest.TensorRTParam(
-            1 << 30, 32, 0, AnalysisConfig.Precision.Int8, False, False)
-        self.fetch_list = [relu_out]
-
-    def append_quantized_op(self, x, param_attr):
-        return fluid.layers.fc(x,
-                               size=100,
-                               num_flatten_dims=3,
-                               param_attr=param_attr,
-                               bias_attr=False,
-                               act=None)
-
-    def set_quant_pattern(self):
-        self.activation_quant_type = 'moving_average_abs_max'
-        self.weight_quant_type = 'abs_max'
-        self.quantized_op_type = 'mul'
-        self.channels = 1
-
-    def test_check_output(self):
-        if core.is_compiled_with_cuda():
-            use_gpu = True
-            self.check_output_with_option(use_gpu, flatten=False, quant=True)
-            self.assertTrue(
-                PassVersionChecker.IsCompatible(
-                    'quant_conv2d_dequant_fuse_pass'))
-
-
-class Quant_Fc2_DequantTest(InferencePassTest):
-    def setUp(self):
-        with fluid.program_guard(self.main_program, self.startup_program):
-            data = fluid.data(
-                name="data", shape=[-1, 3, 32, 32], dtype="float32")
-            param_attr = fluid.ParamAttr(
-                initializer=fluid.initializer.Constant(value=1.0),
-                trainable=False)
-            quantized_op_out = self.append_quantized_op(data, param_attr)
-            relu_out = fluid.layers.relu(quantized_op_out)
-            self.set_quant_pattern()
-
-        self.feeds = {
-            "data": np.random.random([1, 3, 32, 32]).astype("float32"),
-        }
-        self.enable_trt = True
-        self.trt_parameters = Quant_Fc2_DequantTest.TensorRTParam(
-            1 << 30, 32, 0, AnalysisConfig.Precision.Int8, False, False)
-        self.fetch_list = [relu_out]
-
-    def append_quantized_op(self, x, param_attr):
-        return fluid.layers.fc(x,
-                               size=100,
-                               num_flatten_dims=1,
-                               param_attr=param_attr,
-                               bias_attr=False,
-                               act=None)
-
-    def set_quant_pattern(self):
-        self.activation_quant_type = 'moving_average_abs_max'
-        self.weight_quant_type = 'abs_max'
-        self.quantized_op_type = 'mul'
-        self.channels = 1
-
-    def test_check_output(self):
-        if core.is_compiled_with_cuda():
-            use_gpu = True
-            self.check_output_with_option(use_gpu, flatten=False, quant=True)
-            self.assertTrue(
-                PassVersionChecker.IsCompatible(
-                    'quant_conv2d_dequant_fuse_pass'))
-
-
-class Quant_Conv_DequantTest(InferencePassTest):
-    def setUp(self):
-        with fluid.program_guard(self.main_program, self.startup_program):
-            data = fluid.data(
-                name="data", shape=[-1, 3, 32, 32], dtype="float32")
-            param_attr = fluid.ParamAttr(
-                initializer=fluid.initializer.Constant(value=1.0),
-                trainable=False)
-            quantized_op_out = self.append_quantized_op(data, param_attr)
-            relu_out = fluid.layers.relu(quantized_op_out)
-            self.set_quant_pattern()
-
-        self.feeds = {
-            "data": np.random.random([1, 3, 32, 32]).astype("float32"),
-        }
-        self.enable_trt = True
-        self.trt_parameters = Quant_Conv_DequantTest.TensorRTParam(
+        self.trt_parameters = QuantDequantTest.TensorRTParam(
             1 << 30, 32, 0, AnalysisConfig.Precision.Int8, False, False)
         self.fetch_list = [relu_out]
 
@@ -148,10 +62,26 @@ class Quant_Conv_DequantTest(InferencePassTest):
     def test_check_output(self):
         if core.is_compiled_with_cuda():
             use_gpu = True
-            self.check_output_with_option(use_gpu, flatten=False, quant=True)
+            self.check_output_with_option(use_gpu, flatten=True, quant=True)
             self.assertTrue(
                 PassVersionChecker.IsCompatible(
                     'quant_conv2d_dequant_fuse_pass'))
+
+
+class QuantFcDequantTest(QuantDequantTest):
+    def append_quantized_op(self, x, param_attr):
+        return fluid.layers.fc(x,
+                               size=100,
+                               num_flatten_dims=1,
+                               param_attr=param_attr,
+                               bias_attr=False,
+                               act=None)
+
+    def set_quant_pattern(self):
+        self.activation_quant_type = 'moving_average_abs_max'
+        self.weight_quant_type = 'abs_max'
+        self.quantized_op_type = 'mul'
+        self.channels = 1
 
 
 if __name__ == "__main__":
